fix(sana-wm): keep per-frame timestep in fp32 to match NVlabs
User feedback on the §6.13b commit pointed out a hidden precision
variable: ``model_timestep = timestep.to(latents.dtype)`` quantised
the timestep down to bf16 before passing it to
``SanaWmTimestepEmbedder``, while NVlabs' ``LTXFlowEuler.sample``
keeps the timestep in fp32 throughout. The sinusoidal embedding
inside our embedder upcasts to fp32 internally anyway, so casting
to bf16 first only injected truncation error.

Fix: use ``timestep.float()`` when building the per-frame
``(B, 1, F)`` timestep tensor. The legacy scalar path is unchanged.

GPU re-run of the 9-frame reference-alignment harness with all
other settings identical:

| Config | MAE | PSNR | SSIM-Y |
|---|---|---|---|
| bf16 per-frame + main+cam (previous default) | 82.32 | 8.20 | 0.016 |
| **fp32 per-frame + main+cam (new default)**  | **80.74** | **8.34** | 0.020 |
| fp32 per-frame + main-only                   | 79.99 | 8.56 | 0.011 |

-1.58 MAE, +0.14 PSNR for free. Still firmly in the
"weakly-correlated" regime — Break 3 (per-token flow-matching
Euler step) is the main lever for the next chunk of headroom.

+++ vllm_omni/diffusion/models/sana_wm/pipeline_sana_wm.py
@@ -746,7 +746,13 @@ class SanaWmPipeline(
         for timestep in timesteps:
             if use_per_frame_timestep:
                 # (B, 1, F) per-frame timestep, frame 0 forced to 0.
-                model_timestep = timestep.to(latents.dtype).expand(
+                # Keep fp32 to match NVlabs ``LTXFlowEuler.sample`` — the
+                # sinusoidal-embedding inside ``SanaWmTimestepEmbedder``
+                # uses ``timestep.float()`` internally; casting through
+                # the latent dtype (bf16) first would quantise the
+                # timestep before the embed and inject a hidden
+                # precision variable into the contract.
+                model_timestep = timestep.float().expand(
                     cam_batch, 1, cam_frames
                 ).clone()
                 model_timestep[:, :, 0] = 0.0
